requirements.txt
babel==2.16.0
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
docx2pdf==0.1.8
docxcompose==1.4.0
idna==3.6
lxml==5.3.0
numpy==1.26.4
pdfminer.six==20240706
PyAudio==0.2.14
pycparser==2.22
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-docx==1.1.2
pywin32==306
requests==2.31.0
setuptools==69.5.1
six==1.16.0
tk==0.1.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.1
